requirements.txt
Flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
Werkzeug==2.3.7
python-multipart==0.0.6
